Move debounce from hooks to utilities
